feat: removed opÃerator parsing from lexer for more parsing freedom
