+++ compiler/ast/src/lib.rs
@@ -5,3 +5,4 @@
 pub mod tree;
 pub mod ctx;
 pub mod types;
+pub mod operators;
