+++ compiler/lexer/src/lib.rs
@@ -5,4 +5,3 @@
 
 pub mod token;
 pub mod lexer;
-pub mod toks;
