--- compiler/lexer/src/toks/math.rs
@@ -1,10 +1,0 @@
-///! Maths token related utils
-
-/// The different operators
-#[derive(Debug, PartialEq, Clone)]
-pub enum MathOperator {
-	ADD,
-	SUBSTRACT,
-	MULTIPLY,
-	DIVIDE
-}
