--- compiler/lexer/src/toks/mod.rs
@@ -1,2 +1,0 @@
-pub mod math;
-pub mod comp;
